Bump ORM requirements, fix translation walker handling of paginated queries
